# Codechange: Use vehicle viewport hash to find clicked vehicle.