## Motivation / Problem

Clicking on a vehicle in a viewport needs to find the which vehicle was clicked on. This is done by iterating through all vehicles that exist to find the nearest one.

On a game with a lot of vehicles this can take some time (but only in the order of milliseconds.)

<!--
Describe here shortly
* For bug fixes:
    * What problem does this solve?
    * If there is already an issue, link the issue, otherwise describe the problem here.
* For features or gameplay changes:
    * What was the motivation to develop this feature?
    * Does this address any problem with the gameplay or interface?
    * Which group of players do you think would enjoy this feature?
-->


## Description

Use the existing vehicle viewport hash instead.

This avoids having to iterate all vehicles, which can be a minor performance improvement when there are many vehicles.

Performance measurements with the old stalwart wentbourne game:

| version | release      | debug         |
| --------| ------------ | ------------- |
| master  |   avg 1100µs (1.1ms) | avg 6500µs (6.5ms) |
| this pull request | 5-20µs | 10-30µs |

<!--
Describe here shortly
* For bug fixes:
    * How is the problem solved?
* For features or gameplay changes:
    * What does this feature do?
    * How does it improve/solve the situation described under 'motivation'.
-->


## Limitations

The code is more complex as it duplicates the viewport hash loop from `ViewportAddVehicles()`. While that could be abstracted out to allow reuse, it would then be slightly detrimental to viewport drawing performance.

This is a fairly unnecessary performance change, as a user cannot click on vehicles fast enough to notice an issue.

<!--
Describe here
* Is the problem solved in all scenarios?
* Is this feature complete? Are there things that could be added in the future?
* Are there things that are intentionally left out?
* Do you know of a bug or corner case that does not work?
-->


## Checklist for review

Some things are not automated, and forgotten often. This list is a reminder for the reviewers.
* The bug fix is important enough to be backported? (label: 'backport requested')
* This PR touches english.txt or translations? Check the [guidelines](https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md)
* This PR affects the save game format? (label 'savegame upgrade')
* This PR affects the GS/AI API? (label 'needs review: Script API')
    * ai_changelog.hpp, game_changelog.hpp need updating.
    * The compatibility wrappers (compat_*.nut) need updating.
* This PR affects the NewGRF API? (label 'needs review: NewGRF')
    * newgrf_debug_data.h may need updating.
    * [PR must be added to API tracker](https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status)