@PeterN 2024-01-03
> I assume you tested it :)

"It's so quick you don't even see the window appear..."